fix: refactoring of oauth code
