+++ .env.example
@@ -1,5 +1,4 @@
 REACT_APP_API_BASE_URL=
 
-CLIENT_ID=7710a89686dcc822590c
-REDIRECT_URL=http://localhost:1234/
-BACK_END_URL=http://localhost:5000
+REACT_APP_CLIENT_ID=
+REACT_APP_REDIRECT_URI=http://localhost:3000/sign-in
